+++ worlds/pokemon_bw/options/groups.py
@@ -1,0 +1,44 @@
+from Options import OptionGroup
+from . import *
+
+option_groups = [
+    OptionGroup("Encounter", [
+        RandomizeWildPokemon,
+        RandomizeTrainerPokemon,
+        PokemonRandomizationAdjustments,
+        EncounterPlando,
+        WildRandomizationBlacklist,
+        TrainerRandomizationBlacklist,
+    ]),
+    OptionGroup("Stats", [
+        RandomizeBaseStats,
+        RandomizeEvolutions,
+        RandomizeCatchRates,
+        RandomizeLevelUpMovesets,
+        StatsRandomizationAdjustments,
+        StatsPlando,
+    ]),
+    OptionGroup("Items, locations, and progression", [
+        ShuffleBadgeRewards,
+        ShuffleTMRewards,
+        Dexsanity,
+        SeasonControl,
+        ModifyItemPool,
+        ModifyLogic,
+        FillerItemsBlacklist,
+    ]),
+    OptionGroup("Miscellaneous", [
+        AdjustLevels,
+        ModifyLevels,
+        ModifyEncounterRates,
+        ExpMultiplier,
+        AllPokemonSeen,
+        ReplaceEvoMethods,
+        MasterBallSeller,
+        # StartInventoryPool, not allowed in new option group
+        FunnyDialog,
+        PokemonBWTextPlando,
+        PluginOptions,
+        ReusableTMs,
+    ]),
+]